Fix tfs push bug
The fix is ready, which basically:

Handles the case when SCM configuration of a job is empty, which was the
root cause of the previous random TFS push trigger failure.
Add an additional case when a job should be executed – the commit is NOT
null:
a. runPolling() actually checks SCM changes, which in our cases are
none.
b. As long as a commit in linked repro triggers this push (commit is NOT
null), the job should be executed.
Refactors original trigger logic into function “triggerJob”, to avoid
dupe code as these code will be called twice in AbstractHookEvent.java.
I choose not to remove existing code in AbstractHookEvent.java, which
goes through all SCMs and Repos and might still be applied to some old
free style projects. Instead I add the handling of the missing cases
(when SCM configuration of a job is empty) on top of it.

Now if you merge any change to dotnet-linker/enable-ci, all four jobs
starting with “testing” (on link below) will be triggered, and ONLY
those three will be triggered.

https://dotnet-vsts.westus2.cloudapp.azure.com/job/DevDiv_dotnet-linker/job/enable-ci/